ACRFuncs/requirements.txt
azure-functions
requests==2.19.1